+++ packages/studio/src/components/container/audio-panel-container.tsx
@@ -1,0 +1,27 @@
+import { useAudioPanel } from "../../hooks/use-audio-panel";
+import { AudioPanel } from "../element-panel/audio-panel";
+import type { PanelProps } from "../../types";
+
+export const AudioPanelContainer = (props: PanelProps) => {
+  const {
+    items,
+    searchQuery,
+    setSearchQuery,
+    playingAudio,
+    handleSelection,
+    handlePlayPause,
+    handleFileUpload,
+  } = useAudioPanel(props);
+
+  return (
+    <AudioPanel
+      items={items}
+      searchQuery={searchQuery}
+      playingAudio={playingAudio}
+      onSearchChange={setSearchQuery}
+      onItemSelect={handleSelection}
+      onPlayPause={handlePlayPause}
+      onFileUpload={handleFileUpload}
+    />
+  );
+};
