+++ packages/studio/src/components/container/text-panel-container.tsx
@@ -1,0 +1,14 @@
+import { TrackElement } from "@twick/timeline";
+import { TextPanel } from "../element-panel/text-panel";
+import { useTextPanel } from "../../hooks/use-text-panel";
+
+interface TextPanelContainerProps {
+  selectedElement: TrackElement | null;
+  addElement: (element: TrackElement) => void;
+  updateElement: (element: TrackElement) => void;
+}
+
+export function TextPanelContainer(props: TextPanelContainerProps) {
+  const textPanelProps = useTextPanel(props);
+  return <TextPanel {...textPanelProps} />;
+}
