+++ packages/studio/src/components/element-panel/audio-panel.tsx
@@ -1,94 +1,27 @@
-import { useEffect, useState } from "react";
-import { Upload, Wand2, Plus, Volume2, Play, Pause } from "lucide-react";
-import { getMediaManager } from "../../shared";
+import { Wand2, Plus, Volume2, Play, Pause } from "lucide-react";
+import SearchInput from "../shared/search-input";
+import FileInput from "../shared/file-input";
 import type { MediaItem } from "@twick/video-editor";
-import { AudioElement } from "@twick/timeline";
-import SearchInput from "../../shared/search-input";
-import FileInput from "../../shared/file-input";
-import type { PanelProps } from "../../types";
 
-export const AudioPanel = ({ selectedElement, addElement, updateElement }: PanelProps) => {
-  const [playingAudio, setPlayingAudio] = useState<string | null>(null);
-  const [audioElement, setAudioElement] = useState<HTMLAudioElement | null>(
-    null
-  );
+export interface AudioPanelProps {
+  items: MediaItem[];
+  searchQuery: string;
+  playingAudio: string | null;
+  onSearchChange: (query: string) => void;
+  onItemSelect: (item: MediaItem) => void;
+  onPlayPause: (item: MediaItem) => void;
+  onFileUpload: (fileData: { file: File; blobUrl: string }) => void;
+}
 
-  const [items, setItems] = useState<MediaItem[]>([]);
-  const [searchQuery, setSearchQuery] = useState("");
-  const mediaManager = getMediaManager();
-
-  useEffect(() => {
-    const loadItems = async () => {
-      const results = await mediaManager.search({
-        query: searchQuery,
-        type: "audio",
-      });
-      setItems(results);
-    };
-    loadItems();
-  }, [searchQuery]);
-
-  const handleSelection = async (item: MediaItem) => {
-    let audioElement;
-    if(selectedElement instanceof AudioElement) {
-      audioElement = selectedElement;
-      audioElement.setSrc(item.url);
-      await audioElement.updateAudioMeta();
-      updateElement?.(audioElement);
-    } else {
-      audioElement = new AudioElement(item.url);
-      addElement?.(audioElement);
-    }
-  };
-
-  const handlePlayPause = (item: MediaItem) => {
-    if (playingAudio === item.id) {
-      // Stop current audio
-      if (audioElement) {
-        audioElement.pause();
-        audioElement.currentTime = 0;
-      }
-      setPlayingAudio(null);
-      setAudioElement(null);
-    } else {
-      // Stop any currently playing audio
-      if (audioElement) {
-        audioElement.pause();
-        audioElement.currentTime = 0;
-      }
-
-      // Start new audio
-      const audio = new Audio(item.url);
-      audio.play();
-      setPlayingAudio(item.id);
-      setAudioElement(audio);
-
-      // Clean up when audio ends
-      audio.onended = () => {
-        setPlayingAudio(null);
-        setAudioElement(null);
-      };
-    }
-  };
-
-  const handleFileUpload = async (fileData: {
-    file: File;
-    blobUrl: string;
-  }) => {
-    const arrayBuffer = await fileData.file.arrayBuffer();
-    const newItem = await mediaManager.addItem({
-      url: fileData.blobUrl,
-      type: "audio",
-      arrayBuffer,
-      metadata: {
-        name: fileData.file.name,
-        size: fileData.file.size,
-        type: fileData.file.type,
-      },
-    });
-    setItems((prev) => [...prev, newItem]);
-  };
-
+export const AudioPanel = ({
+  items,
+  searchQuery,
+  playingAudio,
+  onSearchChange,
+  onItemSelect,
+  onPlayPause,
+  onFileUpload,
+}: AudioPanelProps) => {
   return (
     <div className="w-72 bg-neutral-800/80 border-r border-gray-600/50 flex flex-col h-full backdrop-blur-md shadow-lg">
       {/* Header */}
@@ -97,34 +30,26 @@ export const AudioPanel = ({ selectedElement, addElement, updateElement }: Panel
 
         {/* Search */}
         <div className="relative mb-3">
-          {/* Search */}
           <SearchInput
             searchQuery={searchQuery}
-            setSearchQuery={setSearchQuery}
+            setSearchQuery={onSearchChange}
           />
-          {/* Upload Button */}
           <FileInput
             id="audio-upload"
             acceptFileTypes={["audio/*"]}
-            onFileLoad={handleFileUpload}
+            onFileLoad={onFileUpload}
             buttonText="Upload"
           />
         </div>
-
-        {/* Upload Button */}
-        <button className="w-full btn btn-primary flex items-center justify-center gap-2 py-2">
-          <Upload className="w-4 h-4" />
-          Upload Audio
-        </button>
       </div>
 
       {/* Audio List */}
       <div className="flex-1 overflow-y-auto p-4">
         <div className="space-y-2">
-          {items.map((item) => (
+          {(items || []).map((item) => (
             <div
               key={item.id}
-              onDoubleClick={() => handleSelection(item)}
+              onDoubleClick={() => onItemSelect(item)}
               className="audio-item group relative cursor-pointer p-3 bg-neutral-700/50 rounded-lg hover:bg-neutral-700/80 transition-all duration-200 border border-transparent hover:border-purple-500/30"
             >
               {/* Audio Info */}
@@ -133,7 +58,7 @@ export const AudioPanel = ({ selectedElement, addElement, updateElement }: Panel
                 <button
                   onClick={(e) => {
                     e.stopPropagation();
-                    handlePlayPause(item);
+                    onPlayPause(item);
                   }}
                   className="w-8 h-8 rounded-full bg-purple-500/80 hover:bg-purple-500 flex items-center justify-center text-white transition-all duration-200 flex-shrink-0"
                 >
@@ -160,7 +85,7 @@ export const AudioPanel = ({ selectedElement, addElement, updateElement }: Panel
                 <button
                   onClick={(e) => {
                     e.stopPropagation();
-                    handleSelection(item);
+                    onItemSelect(item);
                   }}
                   className="w-6 h-6 rounded-full bg-purple-500/60 hover:bg-purple-500 flex items-center justify-center text-white text-xs opacity-0 group-hover:opacity-100 transition-opacity duration-200 flex-shrink-0"
                 >
