+++ packages/studio/src/components/element-panel/circle-panel.tsx
@@ -1,68 +1,38 @@
-import { useEffect, useState } from "react";
 import { inputStyles } from "../../styles/input-styles";
-import { RectElement } from "@twick/timeline";
-import type { PanelProps } from "../../types";
+import type { CirclePanelState, CirclePanelActions } from "../../hooks/use-circle-panel";
 
-const DEFAULT_RECT_PROPS = {
-  cornerRadius: 0,
-  fillColor: "#3b82f6",
-  opacity: 100,
-  strokeColor: "#000000",
-  lineWidth: 0,
-}     
-export function RectPanel({selectedElement, addElement, updateElement}: PanelProps) {
-  const [cornerRadius, setCornerRadius] = useState(DEFAULT_RECT_PROPS.cornerRadius);
-  const [fillColor, setFillColor] = useState(DEFAULT_RECT_PROPS.fillColor);
-  const [opacity, setOpacity] = useState(DEFAULT_RECT_PROPS.opacity);
-  const [strokeColor, setStrokeColor] = useState(DEFAULT_RECT_PROPS.strokeColor);
-  const [lineWidth, setLineWidth] = useState(DEFAULT_RECT_PROPS.lineWidth);
+export type CirclePanelProps = CirclePanelState & CirclePanelActions;
 
-  const handleApplyChanges = () => {
-    let rectElement;
-    if(selectedElement instanceof RectElement) {
-      rectElement = selectedElement;
-      rectElement.setCornerRadius(cornerRadius);
-      rectElement.setOpacity(opacity);
-      rectElement.setStrokeColor(strokeColor);
-      rectElement.setLineWidth(lineWidth);
-      updateElement?.(rectElement);
-    } else {
-      rectElement = new RectElement(fillColor, { width: 200, height: 200 })
-    .setCornerRadius(cornerRadius)
-    .setOpacity(opacity)
-    .setStrokeColor(strokeColor)
-    .setLineWidth(lineWidth);
-    addElement?.(rectElement);
-    }
-  };
-
-  useEffect(() => {
-    if(selectedElement instanceof RectElement) {
-      setCornerRadius(selectedElement.getCornerRadius() ?? DEFAULT_RECT_PROPS.cornerRadius);
-      setFillColor(selectedElement.getFill() ?? DEFAULT_RECT_PROPS.fillColor);
-      setOpacity(selectedElement.getOpacity() ?? DEFAULT_RECT_PROPS.opacity);
-      setStrokeColor(selectedElement.getStrokeColor() ?? DEFAULT_RECT_PROPS.strokeColor);
-      setLineWidth(selectedElement.getLineWidth() ?? DEFAULT_RECT_PROPS.lineWidth);
-    }
-  }, [selectedElement]);
-
+export function CirclePanel({
+  radius,
+  fillColor,
+  opacity,
+  strokeColor,
+  lineWidth,
+  setRadius,
+  setFillColor,
+  setOpacity,
+  setStrokeColor,
+  setLineWidth,
+  handleApplyChanges,
+}: CirclePanelProps) {
   return (
     <div className={inputStyles.panel.container}>
-      <h3 className={inputStyles.panel.title}>Rectangle</h3>
+      <h3 className={inputStyles.panel.title}>Circle</h3>
 
-      {/* Corner Radius */}
+      {/* Radius */}
       <div className={inputStyles.container}>
-        <label className={inputStyles.label.base}>Corner Radius</label>
+        <label className={inputStyles.label.base}>Radius</label>
         <div className={inputStyles.range.container}>
           <input
             type="range"
-            min="0"
-            max="100"
-            value={cornerRadius}
-            onChange={(e) => setCornerRadius(Number(e.target.value))}
+            min="10"
+            max="300"
+            value={radius}
+            onChange={(e) => setRadius(Number(e.target.value))}
             className={inputStyles.range.base}
           />
-          <span className={inputStyles.range.value}>{cornerRadius}px</span>
+          <span className={inputStyles.range.value}>{radius}px</span>
         </div>
       </div>
 
@@ -76,7 +46,7 @@ export function RectPanel({selectedElement, addElement, updateElement}: PanelPro
             onChange={(e) => setFillColor(e.target.value)}
             className={inputStyles.color.picker}
           />
-          <div 
+          <div
             className={inputStyles.color.preview}
             style={{ backgroundColor: fillColor }}
           />
@@ -109,7 +79,7 @@ export function RectPanel({selectedElement, addElement, updateElement}: PanelPro
             onChange={(e) => setStrokeColor(e.target.value)}
             className={inputStyles.color.picker}
           />
-          <div 
+          <div
             className={inputStyles.color.preview}
             style={{ backgroundColor: strokeColor }}
           />
