+++ packages/studio/src/components/element-panel/rect-panel.tsx
@@ -1,72 +1,38 @@
-import { useEffect, useState } from "react";
 import { inputStyles } from "../../styles/input-styles";
-import { CircleElement } from "@twick/timeline";
-import type { PanelProps } from "../../types";
+import type { RectPanelState, RectPanelActions } from "../../hooks/use-rect-panel";
 
-const DEFAULT_CIRCLE_PROPS = {
-  radius: 50,
-  fillColor: "#3b82f6",
-  opacity: 100,
-  strokeColor: "#000000",
-  lineWidth: 0,
-}
-export function CirclePanel({
-  selectedElement,
-  addElement,
-  updateElement,
-}: PanelProps) {
-  const [radius, setRadius] = useState(DEFAULT_CIRCLE_PROPS.radius);
-  const [fillColor, setFillColor] = useState(DEFAULT_CIRCLE_PROPS.fillColor);
-  const [opacity, setOpacity] = useState(DEFAULT_CIRCLE_PROPS.opacity);
-  const [strokeColor, setStrokeColor] = useState(DEFAULT_CIRCLE_PROPS.strokeColor);
-  const [lineWidth, setLineWidth] = useState(DEFAULT_CIRCLE_PROPS.lineWidth);
+export type RectPanelProps = RectPanelState & RectPanelActions;
 
-  const handleApplyChanges = () => {
-    let circleElement;
-    if (selectedElement instanceof CircleElement) {
-      circleElement = selectedElement;
-      circleElement.setRadius(radius);
-      circleElement.setFill(fillColor);
-      circleElement.setOpacity(opacity);
-      circleElement.setStrokeColor(strokeColor);
-      circleElement.setLineWidth(lineWidth);
-      updateElement?.(circleElement);
-    } else {
-      circleElement = new CircleElement(fillColor, radius)
-        .setOpacity(opacity)
-        .setStrokeColor(strokeColor)
-        .setLineWidth(lineWidth);
-      addElement?.(circleElement);
-    }
-  };
-
-  useEffect(() => {
-    if (selectedElement instanceof CircleElement) {
-      setRadius(selectedElement.getRadius() ?? DEFAULT_CIRCLE_PROPS.radius);
-      setFillColor(selectedElement.getFill() ?? DEFAULT_CIRCLE_PROPS.fillColor);
-      setOpacity(selectedElement.getOpacity() ?? DEFAULT_CIRCLE_PROPS.opacity);
-      setStrokeColor(selectedElement.getStrokeColor() ?? DEFAULT_CIRCLE_PROPS.strokeColor);
-      setLineWidth(selectedElement.getLineWidth() ?? DEFAULT_CIRCLE_PROPS.lineWidth);
-    }
-  }, [selectedElement]);
-
+export function RectPanel({
+  cornerRadius,
+  fillColor,
+  opacity,
+  strokeColor,
+  lineWidth,
+  setCornerRadius,
+  setFillColor,
+  setOpacity,
+  setStrokeColor,
+  setLineWidth,
+  handleApplyChanges,
+}: RectPanelProps) {
   return (
     <div className={inputStyles.panel.container}>
-      <h3 className={inputStyles.panel.title}>Circle</h3>
+      <h3 className={inputStyles.panel.title}>Rectangle</h3>
 
-      {/* Radius */}
+      {/* Corner Radius */}
       <div className={inputStyles.container}>
-        <label className={inputStyles.label.base}>Radius</label>
+        <label className={inputStyles.label.base}>Corner Radius</label>
         <div className={inputStyles.range.container}>
           <input
             type="range"
-            min="10"
-            max="300"
-            value={radius}
-            onChange={(e) => setRadius(Number(e.target.value))}
+            min="0"
+            max="100"
+            value={cornerRadius}
+            onChange={(e) => setCornerRadius(Number(e.target.value))}
             className={inputStyles.range.base}
           />
-          <span className={inputStyles.range.value}>{radius}px</span>
+          <span className={inputStyles.range.value}>{cornerRadius}px</span>
         </div>
       </div>
 
@@ -80,7 +46,7 @@ export function CirclePanel({
             onChange={(e) => setFillColor(e.target.value)}
             className={inputStyles.color.picker}
           />
-          <div
+          <div 
             className={inputStyles.color.preview}
             style={{ backgroundColor: fillColor }}
           />
@@ -113,7 +79,7 @@ export function CirclePanel({
             onChange={(e) => setStrokeColor(e.target.value)}
             className={inputStyles.color.picker}
           />
-          <div
+          <div 
             className={inputStyles.color.preview}
             style={{ backgroundColor: strokeColor }}
           />
