+++ packages/studio/src/components/element-panel/subtitles-panel.tsx
