+++ packages/studio/src/components/properties/element-props.tsx
@@ -1,0 +1,100 @@
+import { Settings } from "lucide-react";
+import { AccordionItem } from "../shared/accordion-item";
+
+interface ElementPropsProps {
+  isOpen: boolean;
+  onToggle: () => void;
+}
+
+export function ElementProps({ isOpen, onToggle }: ElementPropsProps) {
+  return (
+    <AccordionItem
+      title="All Properties"
+      icon={<Settings className="w-4 h-4" />}
+      isOpen={isOpen}
+      onToggle={onToggle}
+    >
+      <div className="space-y-3">
+        {/* Position & Size */}
+        <div className="bg-neutral-800/40 rounded-lg p-2.5 border border-gray-600/20">
+          <h5 className="text-xs font-semibold text-gray-200 mb-2 flex items-center gap-1.5">
+            <div className="w-1.5 h-1.5 bg-purple-400 rounded-full"></div>
+            Position & Size
+          </h5>
+          <div className="grid grid-cols-2 gap-2">
+            <div>
+              <label className="block text-xs text-gray-400 mb-1">X</label>
+              <input
+                type="number"
+                defaultValue="0"
+                className="w-full bg-neutral-700/60 border border-gray-600/40 rounded-md text-white text-xs px-2 py-1.5 focus:border-purple-500 focus:outline-none focus:ring-1 focus:ring-purple-500/30 transition-all duration-200"
+              />
+            </div>
+            <div>
+              <label className="block text-xs text-gray-400 mb-1">Y</label>
+              <input
+                type="number"
+                defaultValue="0"
+                className="w-full bg-neutral-700/60 border border-gray-600/40 rounded-md text-white text-xs px-2 py-1.5 focus:border-purple-500 focus:outline-none focus:ring-1 focus:ring-purple-500/30 transition-all duration-200"
+              />
+            </div>
+            <div>
+              <label className="block text-xs text-gray-400 mb-1">Width</label>
+              <input
+                type="number"
+                defaultValue="100"
+                className="w-full bg-neutral-700/60 border border-gray-600/40 rounded-md text-white text-xs px-2 py-1.5 focus:border-purple-500 focus:outline-none focus:ring-1 focus:ring-purple-500/30 transition-all duration-200"
+              />
+            </div>
+            <div>
+              <label className="block text-xs text-gray-400 mb-1">Height</label>
+              <input
+                type="number"
+                defaultValue="100"
+                className="w-full bg-neutral-700/60 border border-gray-600/40 rounded-md text-white text-xs px-2 py-1.5 focus:border-purple-500 focus:outline-none focus:ring-1 focus:ring-purple-500/30 transition-all duration-200"
+              />
+            </div>
+          </div>
+        </div>
+
+        {/* Opacity */}
+        <div className="bg-neutral-800/40 rounded-lg p-2.5 border border-gray-600/20">
+          <h5 className="text-xs font-semibold text-gray-200 mb-2 flex items-center gap-1.5">
+            <div className="w-1.5 h-1.5 bg-purple-400 rounded-full"></div>
+            Opacity
+          </h5>
+          <input
+            type="range"
+            min="0"
+            max="100"
+            defaultValue="100"
+            className="w-full h-1.5 bg-gradient-to-r from-purple-500/30 to-neutral-600/50 rounded-full appearance-none cursor-pointer slider-thumb"
+          />
+          <div className="flex justify-between text-xs text-gray-400 mt-1">
+            <span>0%</span>
+            <span>100%</span>
+          </div>
+        </div>
+
+        {/* Rotation */}
+        <div className="bg-neutral-800/40 rounded-lg p-2.5 border border-gray-600/20">
+          <h5 className="text-xs font-semibold text-gray-200 mb-2 flex items-center gap-1.5">
+            <div className="w-1.5 h-1.5 bg-purple-400 rounded-full"></div>
+            Rotation
+          </h5>
+          <input
+            type="range"
+            min="0"
+            max="360"
+            defaultValue="0"
+            className="w-full h-1.5 bg-gradient-to-r from-purple-500/30 to-neutral-600/50 rounded-full appearance-none cursor-pointer slider-thumb"
+          />
+          <div className="flex justify-between text-xs text-gray-400 mt-1">
+            <span>0°</span>
+            <span>360°</span>
+          </div>
+        </div>
+      </div>
+    </AccordionItem>
+  );
+}
