+++ packages/studio/src/components/shared/color-input.tsx
