+++ packages/studio/src/components/shared/media-manager.ts
