+++ packages/studio/src/components/shared/search-input.tsx
