+++ packages/studio/src/hooks/use-studio-manager.tsx
